Added deregister command
This command will only remove the profile in the databse that is assosiated with the invoking user's DiscordID. It will not remove any characters. Will have to think about whether this is a good or a bad choice... Ultimetly it might just be enough for this command to check if the user has any characters registered to themselves before letting the user proceed to remove their profile. This stops them from removing ALL of their characters by just sending one command... It's something to think about